Now accepts one or more package names on command line; detect existing packages with only *.py[co] files

+++ addsitepackage2venv.py
@@ -1,19 +1,21 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-"""Make a Python module or package installed in one of the system-wide
-site-packages directories available to the currently active virtual environment
-by sym-linking it into the virtualenv's site-packages directory.
+"""Make one ore more Python modules or packages installed in one of the
+system-wide site-packages directories available in the currently active
+virtual environment by sym-linking it into the virtualenv's site-packages
+directory.
 
 """
 
 
 __author__ = "Christopher Arndt"
-__version__ = "0.2 ($Rev$)"
+__version__ = "0.3 ($Rev$)"
 __license__ = "MIT License"
 __usage__ = "Usage: addsitepackage2venv <package>"
 
 import argparse
 import distutils.sysconfig
+import glob
 import os
 import site
 import sys
@@ -24,7 +26,7 @@ from os.path import basename, dirname, exists, isdir, islink, join
 MODULE_EXTENSIONS = ['', '.pth', '.py', '.pyc', '.pyo']
 
 if sys.platform.startswith('win'):
-    MODULE_EXTENSIONS += ['.dll', '.pyd']
+    MODULE_EXTENSIONS += ['.dll', '.pyd', '.pyw']
 else:
     MODULE_EXTENSIONS += ['.so']
 
@@ -35,8 +37,8 @@ def main(args=None):
         action="store_true", help="Overwrite existing symlinks")
     argparser.add_argument('-v', '--verbose',
         action="store_true", help="Be more verbose")
-    argparser.add_argument('package',
-        help="Name of system package to add to virtual environment.")
+    argparser.add_argument('packages', nargs='+',
+        help="Name(s) of system package(s) to add to virtual environment.")
 
     args = argparser.parse_args(args if args is not None else sys.argv[1:])
 
@@ -54,58 +56,66 @@ def main(args=None):
             "Nothing to do.")
         return 1
 
-    venvpackage = join(venvsitedir, args.package)
+    for package in args.packages:
+        venvpackage = join(venvsitedir, package)
 
-    for ext in MODULE_EXTENSIONS:
-        venvmodule = venvpackage + ext
+        for ext in MODULE_EXTENSIONS:
+            venvmodule = venvpackage + ext
 
-        if ((ext == '' and exists(join(venvpackage, '__init__.py'))) or
-                exists(venvmodule)):
-            print("Package %s is already in installed in '%s' as '%s'." %
-                (args.package, venvsitedir, basename(venvpackage + ext)))
+            if ((ext == '' and glob.glob(join(venvpackage, '__init__.py*'))) or
+                    exists(venvmodule)):
+                print("Package %s is already in installed in '%s' as '%s'." %
+                    (package, venvsitedir, basename(venvpackage + ext)))
 
-            if islink(venvpackage + ext):
-                print("Symlinks to: %s" % os.readlink(venvpackage + ext))
+                if islink(venvpackage + ext):
+                    print("Symlinks to: %s" % os.readlink(venvpackage + ext))
 
-            if not args.force:
-                return 1
+                if not args.force:
+                    return 1
 
     site.virtual_addsitepackages(set(sys.path))
 
-    for sitedir in sys.path:
-        if not sitedir or not isdir(sitedir) or sitedir.startswith(venv):
-            continue
+    for package in args.packages:
+        venvpackage = join(venvsitedir, package)
 
         if args.verbose:
-            print("Checking directory '%s'..." % sitedir)
+            print("Checking for package '%s'..." % package)
 
-        pkgdir = join(sitedir, args.package)
+        for sitedir in sys.path:
+            if (not sitedir or not isdir(sitedir) or sitedir.startswith(venv)
+                    or sitedir == dirname(sys.argv[0])):
+                continue
 
-        found = False
-        for ext in MODULE_EXTENSIONS:
-            module = pkgdir + ext
+            if args.verbose:
+                print("Checking directory '%s'..." % sitedir)
 
-            if ((ext == '' and exists(join(pkgdir, '__init__.py'))) or
-                    exists(module)):
-                try:
-                    if exists(venvpackage + ext) and args.force:
-                        os.unlink(venvpackage + ext)
+            pkgdir = join(sitedir, package)
 
-                    os.symlink(module, venvpackage + ext)
-                    print("Created symlink: %s -> %s" %
-                        (module, venvpackage + ext))
-                except (IOError, OSError) as exc:
-                    print("Could not create symlink %s -> %s: %s" %
-                        (module, venvpackage + ext, exc))
-                found = True
+            found = False
+            for ext in MODULE_EXTENSIONS:
+                module = pkgdir + ext
 
-        if found:
-            break
+                if ((ext == '' and glob.glob(join(pkgdir, '__init__.py*'))) or
+                        exists(module)):
+                    try:
+                        if exists(venvpackage + ext) and args.force:
+                            os.unlink(venvpackage + ext)
 
-    else:
-        print("No package '%s' found in site-packages directories." %
-            args.package)
-        return 1
+                        os.symlink(module, venvpackage + ext)
+                        print("Created symlink: %s -> %s" %
+                            (module, venvpackage + ext))
+                    except (IOError, OSError) as exc:
+                        print("Could not create symlink %s -> %s: %s" %
+                            (module, venvpackage + ext, exc))
+                    found = True
+
+            if found:
+                break
+
+        else:
+            print("No package '%s' found in site-packages directories." %
+                package)
+            return 1
 
 
 if __name__ == '__main__':
